Revert "Feature/serialization"
