--- FLG.Cs.Serialization/ISerializable.cs
@@ -1,6 +1,0 @@
-﻿namespace FLG.Cs.Serialization {
-    public interface ISerializable {
-        public void Serialize();
-        public void Deserialize();
-    }
-}
