+++ FLG.Framework.sln
@@ -23,12 +23,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FLG.Cs.Logger", "FLG.Cs.Log
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FLG.Cs.Factory", "FLG.Cs.Factory\FLG.Cs.Factory.csproj", "{262A7F4B-DF3B-450D-A24A-E6185628B446}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FLG.Cs.Serialization", "FLG.Cs.Serialization\FLG.Cs.Serialization.csproj", "{15C0EF46-1EDE-4014-96E4-5BCC528EEB44}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FLG.Cs.IO", "FLG.Cs.IO\FLG.Cs.IO.csproj", "{CEA31421-CB04-4A19-8635-5E7F5D78CA24}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FLG.Cs.Serialization.Tests", "FLG.Cs.Serialization.Tests\FLG.Cs.Serialization.Tests.csproj", "{2E69AB9A-0953-4330-B427-782AB9BE7AD2}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -67,18 +61,6 @@ Global
 		{262A7F4B-DF3B-450D-A24A-E6185628B446}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{262A7F4B-DF3B-450D-A24A-E6185628B446}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{262A7F4B-DF3B-450D-A24A-E6185628B446}.Release|Any CPU.Build.0 = Release|Any CPU
-		{15C0EF46-1EDE-4014-96E4-5BCC528EEB44}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{15C0EF46-1EDE-4014-96E4-5BCC528EEB44}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{15C0EF46-1EDE-4014-96E4-5BCC528EEB44}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{15C0EF46-1EDE-4014-96E4-5BCC528EEB44}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CEA31421-CB04-4A19-8635-5E7F5D78CA24}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CEA31421-CB04-4A19-8635-5E7F5D78CA24}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CEA31421-CB04-4A19-8635-5E7F5D78CA24}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CEA31421-CB04-4A19-8635-5E7F5D78CA24}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2E69AB9A-0953-4330-B427-782AB9BE7AD2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2E69AB9A-0953-4330-B427-782AB9BE7AD2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2E69AB9A-0953-4330-B427-782AB9BE7AD2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2E69AB9A-0953-4330-B427-782AB9BE7AD2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -93,9 +75,6 @@ Global
 		{E64A9578-D65C-42B3-B195-F699EBBA4295} = {ADB89E8B-CF3B-4795-9F54-29447BF9662E}
 		{15016A19-6F82-4D77-A0DF-0041929FB4B8} = {ADB89E8B-CF3B-4795-9F54-29447BF9662E}
 		{262A7F4B-DF3B-450D-A24A-E6185628B446} = {ADB89E8B-CF3B-4795-9F54-29447BF9662E}
-		{15C0EF46-1EDE-4014-96E4-5BCC528EEB44} = {ADB89E8B-CF3B-4795-9F54-29447BF9662E}
-		{CEA31421-CB04-4A19-8635-5E7F5D78CA24} = {ADB89E8B-CF3B-4795-9F54-29447BF9662E}
-		{2E69AB9A-0953-4330-B427-782AB9BE7AD2} = {59DF2940-1767-4671-8EC3-35A915287FAC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {246693A2-ECDB-4309-9A0D-CDB090114039}
